Add ButtonProps and Size types: Introduce new ButtonProps type for button components and a Size type for consistent sizing across UI elements. Update Button and IconButton components to utilize the new types, and remove the obsolete UnlockIcon component for cleaner codebase.
