+++ lib/ui/buttons/IconButton/index.tsx
@@ -1,12 +1,14 @@
 import { UnstyledButton } from '@lib/ui/buttons/UnstyledButton'
 import { horizontalPadding } from '@lib/ui/css/horizontalPadding'
 import { Spinner } from '@lib/ui/loaders/Spinner'
-import { ButtonProps, Size } from '@lib/ui/props'
 import { getColor } from '@lib/ui/theme/getters'
 import { Tooltip } from '@lib/ui/tooltips/Tooltip'
 import { match } from '@lib/utils/match'
 import { FC } from 'react'
 import styled, { css } from 'styled-components'
+
+import { Size } from '../../core/Size'
+import { ButtonProps } from '../ButtonProps'
 
 type ButtonSize = Extract<Size, 'xs' | 'sm' | 'md' | 'lg' | 'xl'>
