+++ lib/ui/core/Size.tsx
@@ -1,0 +1,1 @@
+export type Size = 'xs' | 'sm' | 'md' | 'lg' | 'xl' | 'xxl'
