+++ lib/ui/css/takeWholeSpaceAbsolutely.tsx
@@ -2,7 +2,7 @@ import styled, { css } from 'styled-components'
 
 import { takeWholeSpace } from './takeWholeSpace'
 
-export const takeWholeSpaceAbsolutely = css`
+const takeWholeSpaceAbsolutely = css`
   position: absolute;
   left: 0;
   top: 0;
